Use real width and height
